Split endpoints into separate files
