+++ src/terraformCloud/index.ts
@@ -3,254 +3,18 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-import { Zodios, ZodiosPlugin, makeApi, makeParameters } from '@zodios/core';
-import { z } from 'zod';
+import * as vscode from 'vscode';
+import { Zodios, ZodiosPlugin } from '@zodios/core';
 import { pluginToken, pluginHeader } from '@zodios/plugins';
 import { TerraformCloudAuthenticationProvider } from '../providers/authenticationProvider';
-import * as vscode from 'vscode';
+import { accountEndpoints } from './account';
+import { organizationEndpoints } from './organization';
+import { projectEndpoints } from './project';
+import { runEndpoints } from './run';
+import { workspaceEndpoints } from './workspace';
 
 const defaultHostname = 'app.terraform.io';
 const basePath = '/api/v2';
-
-const paginationParams = makeParameters([
-  {
-    name: 'page[number]',
-    type: 'Query',
-    description: 'Page number',
-    schema: z.number().positive().optional(),
-  },
-  {
-    name: 'page[size]',
-    type: 'Query',
-    description: 'Count of records',
-    schema: z.number().positive().optional(),
-  },
-]);
-
-const projectFilterParams = makeParameters([
-  {
-    name: 'filter[project][id]',
-    type: 'Query',
-    description: 'Restricts results to workspaces in the specific project',
-    schema: z.string().optional(),
-  },
-]);
-
-const paginationMeta = z.object({
-  'current-page': z.number(),
-  'page-size': z.number(),
-  'next-page': z.number().nullable(),
-  'prev-page': z.number().nullable(),
-  'total-count': z.number(),
-  'total-pages': z.number(),
-});
-
-const accountDetails = z.object({
-  data: z.object({
-    id: z.string(),
-    type: z.string(),
-    attributes: z.object({
-      username: z.string(),
-      email: z.string(),
-    }),
-  }),
-});
-
-const accountEndpoints = makeApi([
-  {
-    method: 'get',
-    path: '/account/details',
-    alias: 'getUser',
-    description: 'Get user details',
-    response: accountDetails,
-  },
-]);
-
-const organizations = z.object({
-  data: z.array(
-    z.object({
-      id: z.string(),
-      attributes: z.object({
-        'external-id': z.string(),
-        name: z.string(),
-      }),
-    }),
-  ),
-  meta: z
-    .object({
-      pagination: paginationMeta.optional(),
-    })
-    .optional(),
-});
-
-const organizationMemberships = z.object({
-  data: z.array(
-    z.object({
-      id: z.string(),
-      attributes: z.object({
-        status: z.enum(['active', 'invited']),
-      }),
-      relationships: z.object({
-        organization: z.object({
-          data: z.object({
-            id: z.string(),
-          }),
-        }),
-      }),
-    }),
-  ),
-});
-
-const organizationEndpoints = makeApi([
-  {
-    method: 'get',
-    path: '/organizations',
-    alias: 'listOrganizations',
-    description: 'List organizations of the current user',
-    response: organizations,
-    parameters: paginationParams,
-  },
-  {
-    method: 'get',
-    path: '/organization-memberships',
-    alias: 'listOrganizationMemberships',
-    description: 'List organization memberships of the current user',
-    response: organizationMemberships,
-  },
-]);
-
-const project = z.object({
-  id: z.string(),
-  attributes: z.object({
-    name: z.string(),
-  }),
-});
-
-const projects = z.object({
-  data: z.array(project),
-  meta: z.object({
-    pagination: paginationMeta,
-  }),
-});
-
-const projectEndpoints = makeApi([
-  {
-    method: 'get',
-    path: '/organizations/:organization_name/projects',
-    alias: 'listProjects',
-    description: 'List projects in the organization',
-    response: projects,
-    parameters: paginationParams,
-  },
-  {
-    method: 'get',
-    path: '/projects/:project_id',
-    alias: 'getProject',
-    description: 'Get details on a project',
-    response: z.object({
-      data: project,
-    }),
-  },
-]);
-
-const workspace = z.object({
-  data: z.object({
-    description: z.string(),
-    environment: z.string(),
-    'execution-mode': z.enum(['']),
-    name: z.string(),
-    source: z.string(),
-    'updated-at': z.date(),
-    'run-failures': z.number(),
-  }),
-});
-
-const workspaces = z.object({
-  data: z.array(
-    z.object({
-      id: z.string(),
-      attributes: z.object({
-        description: z.string(),
-        environment: z.string(),
-        'execution-mode': z.enum(['']),
-        name: z.string(),
-        source: z.string(),
-        'updated-at': z.date(),
-        'run-failures': z.number(),
-      }),
-    }),
-  ),
-  meta: z.object({
-    pagination: paginationMeta,
-  }),
-});
-
-const workspaceEndpoints = makeApi([
-  {
-    method: 'get',
-    path: '/organizations/:organization_name/workspaces',
-    alias: 'listWorkspaces',
-    description: 'List workspaces in the organization',
-    response: workspaces,
-    parameters: [...paginationParams, ...projectFilterParams],
-  },
-  {
-    method: 'get',
-    path: '/workspaces/:workspace_id',
-    alias: 'getWorkspace',
-    description: 'Get details on a workspace',
-    response: workspace,
-  },
-]);
-
-const run = z.object({
-  data: z.object({
-    id: z.string(),
-    attributes: z.object({
-      'created-at': z.date(),
-      message: z.string(),
-      source: z.string(),
-      status: z.enum(['']),
-      'trigger-reason': z.enum(['']),
-    }),
-  }),
-});
-
-const runs = z.object({
-  data: z.array(
-    z.object({
-      id: z.string(),
-      attributes: z.object({
-        'created-at': z.date(),
-        message: z.string(),
-        source: z.string(),
-        status: z.enum(['']),
-        'trigger-reason': z.enum(['']),
-      }),
-    }),
-  ),
-  meta: z.object({
-    pagination: paginationMeta,
-  }),
-});
-
-const runEndpoints = makeApi([
-  {
-    method: 'get',
-    path: '/workspaces/:workspace_id/runs',
-    alias: 'listRuns',
-    description: 'List Runs in a Workspace',
-    response: runs,
-    parameters: paginationParams,
-  },
-  {
-    method: 'get',
-    path: '/runs/:run_id',
-    alias: 'getRun',
-    description: 'Show details of a specific run',
-    response: run,
-  },
-]);
 
 const jsonHeader = pluginHeader('Content-Type', async () => 'application/vnd.api+json');
